refactor: 移除废弃宏并消除 unsafe 代码
- 删除未使用的 getter/setter 宏（src/utils/macros.rs），简化 utils 模块
- 负载采样改用 AtomicI64 替代 static mut，去除 unsafe 读写
- 为 main.rs / inotify.rs 的 unsafe 块补充 SAFETY 注释，unwrap 改为 expect
- 前台应用包名正则改为 Lazy 预编译，避免每次调用重复编译
- get_config_list / get_v2_supported_freqs 返回切片引用，消除 clone 开销
- margin 字段由 u32 统一为 i64，移除 try_into().unwrap()
- 修复启动时序：频率/电压初始化提前至监控线程启动前，保证状态一致
- 激进降频跳过防抖延迟立即降频；空闲状态改为私有字段 + set_idle 访问器

+++ src/model/frequency_strategy.rs
@@ -6,7 +6,7 @@ pub struct FrequencyStrategy {
     /// 降频防抖时间
     pub down_debounce_time: u64, // 降频防抖时间（毫秒）
     /// 调整余量
-    pub margin: u32, // 频率调整余量（MHz）
+    pub margin: i64, // 频率调整余量（MHz）
     /// 激进降频开关
     pub aggressive_down: bool, // 是否启用激进降频
     /// 采样间隔
@@ -28,7 +28,7 @@ impl FrequencyStrategy {
     }
 
     /// 设置频率调整余量
-    pub fn set_margin(&mut self, margin: u32) {
+    pub fn set_margin(&mut self, margin: i64) {
         self.margin = margin;
     }
 
@@ -48,7 +48,7 @@ impl FrequencyStrategy {
     }
 
     /// 获取余量
-    pub fn get_margin(&self) -> u32 {
+    pub fn get_margin(&self) -> i64 {
         self.margin
     }
